requirements.txt
python-telegram-bot==20.3
telethon==1.30.3
requests